♻️ remove parking_config.py
