+++ tactics2d/physics/single_track_kinematics.py
@@ -56,16 +56,7 @@ class SingleTrackKinematics(PhysicsModelBase):
         self.accel_range = accel_range
         self.delta_t = min(delta_t, MAX_DELTA_T)
 
-    def _step(
-        self,
-        x: float,
-        y: float,
-        heading: float,
-        speed: float,
-        accel: float,
-        steer: float,
-        dt: float,
-    ):
+    def _step(self, x, y, heading, speed, accel, steer, dt):
         new_x = x + speed * np.cos(heading) * dt
         new_y = y + speed * np.sin(heading) * dt
 
@@ -124,9 +115,8 @@ class SingleTrackKinematics(PhysicsModelBase):
                 step - (dt - self.delta_t),
             )
         # recover the geometry center from rear axle center
-        x, y = rear_center_x + self.dist_rear_hang * np.cos(
-            heading
-        ), rear_center_y + self.dist_rear_hang * np.sin(heading)
+        x = rear_center_x + self.dist_rear_hang * np.cos(heading)
+        y = rear_center_y + self.dist_rear_hang * np.sin(heading)
         new_state = State(
             state.frame + int(step * 1000),
             x=x,
